feat(reservation-form): agregar componentes para selección de habitación
