+++ src/app/(admin)/reservation/_components/create/CreateReservationForm.tsx
@@ -1,6 +1,6 @@
-import React, { useCallback, useEffect, useRef, useState } from "react";
-import { Check, ChevronsUpDown, ListCheck, MapPinHouse, Trash2, UserRoundCheck } from "lucide-react";
-import { Controller, UseFieldArrayReturn, UseFormReturn } from "react-hook-form";
+import React, { useEffect, useRef, useState } from "react";
+import { ListCheck, MapPinHouse, UserRoundCheck } from "lucide-react";
+import { UseFieldArrayReturn, UseFormReturn } from "react-hook-form";
 
 import { Customer } from "@/app/(admin)/customers/_types/customer";
 import { CustomFormDescription } from "@/components/form/CustomFormDescription";
@@ -8,32 +8,24 @@ import ErrorMessageForm from "@/components/form/ErrorMessageForm";
 import LoadingFormSkeleton from "@/components/form/LoadingFormSkeleton";
 import { TextareaWithIcon } from "@/components/form/TextareaWithIcon";
 import { InputWithIcon } from "@/components/input-with-icon";
-import { Button } from "@/components/ui/button";
-import { Command, CommandEmpty, CommandGroup, CommandInput, CommandItem, CommandList } from "@/components/ui/command";
 import { Form, FormControl, FormDescription, FormField, FormItem, FormLabel, FormMessage } from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
-import { PhoneInput } from "@/components/ui/phone-input";
-import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover";
 import { Select, SelectContent, SelectGroup, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select";
 import { Separator } from "@/components/ui/separator";
 import { Switch } from "@/components/ui/switch";
-import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@/components/ui/table";
-import { Textarea } from "@/components/ui/textarea";
 import { cn } from "@/lib/utils";
 import { SelectOption } from "@/types/form/select-option";
 import { useAllAvailableRoomsInTimeInterval } from "../../_hooks/use-roomAvailability";
 import {
   CreateReservationInput,
   DetailedRoom,
-  DocumentType,
   reservationStatusSelectOptions,
 } from "../../_schemas/reservation.schemas";
 import { FORMSTATICS } from "../../_statics/forms";
-import { documentTypeStatusConfig } from "../../_types/document-type.enum.config";
 import { reservationStatusConfig } from "../../_types/reservation-enum.config";
 import { GenericAvailabilityParams } from "../../_types/room-availability-query-params";
-import { SearchCustomerCombobox } from "../search/SearchCustomerCombobox";
 import BookingCalendarTime from "./BookingCalendarTime";
+import CreateHeaderReservation from "./CreateHeaderReservation";
+import CreateReservationGuestTable from "./CreateReservationGuestTable";
 
 interface CreateReservationFormProps extends Omit<React.ComponentPropsWithRef<"form">, "onSubmit"> {
   children: React.ReactNode;
@@ -48,11 +40,12 @@ export default function CreateReservationForm({
   onSubmit,
   controlledFieldArray,
 }: CreateReservationFormProps) {
-  // const { dataRoomsAll } = useRooms();
   const [allowGuests, setAllowGuests] = useState(true);
   const [selectedRoom, setSelectedRoom] = useState<DetailedRoom | undefined>(undefined);
   const [guestNumber, setGuestNumber] = useState<number>(0);
   const [isRoomAvailable, setIsRoomAvailable] = useState(true);
+
+  const initialLoadCompleted = useRef(false);
 
   const defaultCheckInCheckOutDates: GenericAvailabilityParams = {
     checkInDate: form.getValues("checkInDate"),
@@ -61,18 +54,7 @@ export default function CreateReservationForm({
   const { isLoading, isError, error, availableRooms, checkAvailability, refetch } =
     useAllAvailableRoomsInTimeInterval(defaultCheckInCheckOutDates);
   const { watch, register } = form;
-  const { fields, append, remove } = controlledFieldArray;
-  const watchFieldArray = watch("guests");
-  const controlledFields = fields.map((field, index) => {
-    const watchItem = watchFieldArray?.[index];
-    return {
-      ...field,
-      ...(watchItem ?? {}),
-    };
-  });
-
-  const defaultOriginalCheckInDate = useRef(form.getValues("checkInDate"));
-  const defaultOriginalCheckOutDate = useRef(form.getValues("checkOutDate"));
+  const { append, remove } = controlledFieldArray;
 
   const checkInDate = watch("checkInDate");
   const checkOutDate = watch("checkOutDate");
@@ -87,57 +69,48 @@ export default function CreateReservationForm({
     setSelectedRoom(room);
   };
 
-  // 1. Referencia para almacenar el último valor verificado
-  const lastCheckedRef = useRef({
-    roomId: "",
-    checkInDate: defaultOriginalCheckInDate.current,
-    checkOutDate: defaultOriginalCheckOutDate.current,
+  // Referencia para controlar las llamadas a verificación de disponibilidad
+  const availabilityCheckRef = useRef({
+    lastCheckInDate: "",
+    lastCheckOutDate: "",
+    isChecking: false,
   });
 
-  // 2. Estado para rastrear si está en proceso de verificación
-  const [isChecking, setIsChecking] = useState(false);
+  // Ejecutar solo cuando cambian las fechas relevantes
+  useEffect(() => {
+    // Solo verificar disponibilidad cuando ambos valores estén presentes
+    if (checkInDate && checkOutDate) {
+      // Verificar si ya hemos procesado esta combinación de fechas
+      if (
+        availabilityCheckRef.current.lastCheckInDate === checkInDate &&
+        availabilityCheckRef.current.lastCheckOutDate === checkOutDate
+      ) {
+        // Ya hemos procesado esta combinación, no hacer nada
+        return;
+      }
 
-  // 3. Memoizar la función de verificación de disponibilidad
-  const memoizedCheckAvailability = useCallback(() => {
-    // // Solo verificar si hay un ID de habitación
-    // if (!roomId) return;
+      // Actualizar referencias antes de la verificación
+      availabilityCheckRef.current.lastCheckInDate = checkInDate;
+      availabilityCheckRef.current.lastCheckOutDate = checkOutDate;
+      availabilityCheckRef.current.isChecking = true;
 
-    // Si ya está verificando, no iniciar otra verificación
-    if (isChecking) return;
-
-    // Comparar con valores anteriores para evitar verificaciones redundantes
-    if (
-      lastCheckedRef.current.roomId === roomId &&
-      lastCheckedRef.current.checkInDate === checkInDate &&
-      lastCheckedRef.current.checkOutDate === checkOutDate
-    ) {
-      return;
-    }
-
-    // Marcar como verificando
-    setIsChecking(true);
-
-    // Actualizar referencia
-    lastCheckedRef.current = {
-      roomId,
-      checkInDate,
-      checkOutDate,
-    };
-
-    // Realizar la verificación después de un pequeño retraso
-    setTimeout(() => {
+      // Realizar la verificación
       checkAvailability({
         checkInDate,
         checkOutDate,
       });
-      setIsChecking(false);
-    }, 300);
-  }, [roomId, checkInDate, checkOutDate, isChecking, checkAvailability]);
 
-  // 4. Efecto que ejecuta la verificación cuando cambian los valores importantes
+      // Marcar como completada la verificación
+      availabilityCheckRef.current.isChecking = false;
+    }
+  }, [checkInDate, checkOutDate]);
+
+  // --- Efecto para marcar cuando la carga inicial se ha completado ---
   useEffect(() => {
-    memoizedCheckAvailability();
-  }, [memoizedCheckAvailability]);
+    if (!isLoading && !initialLoadCompleted.current) {
+      initialLoadCompleted.current = true;
+    }
+  }, [isLoading]);
 
   const handleAddGuest = () => {
     append({
@@ -173,95 +146,22 @@ export default function CreateReservationForm({
       };
     }) ?? [];
 
-  if (isLoading) {
-    return <LoadingFormSkeleton></LoadingFormSkeleton>;
+  if (isLoading && !initialLoadCompleted.current) {
+    return <LoadingFormSkeleton />;
   }
+
   if (isError) {
     return <ErrorMessageForm error={error} refetch={refetch}></ErrorMessageForm>;
   }
-  // if (roomOptions.length === 0) {
-  //   const error = new Error("No hay habitaciones disponibles para las fechas seleccionadas por defecto");
-  //   return <ErrorMessageForm
-  //   error={error} refetch={refetch}></ErrorMessageForm>;
-  // }
-
   return (
     <Form {...form}>
       <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4 grid grid-cols-1 sm:grid-cols-2 gap-4">
-        <FormItem className="w-full col-span-2 sm:col-span-1">
-          <FormLabel>{FORMSTATICS.customerId.label}</FormLabel>
-          {/* <SearchCustomerCombobox onValueChange={onSearchCustomerFound} /> */}
-          <FormControl>
-            <SearchCustomerCombobox onValueChange={onSearchCustomerFound} className="max-w-none" />
-          </FormControl>
-          <CustomFormDescription
-            required={FORMSTATICS.customerId.required}
-            validateOptionalField={true}
-          ></CustomFormDescription>
-          <FormMessage>
-            {form.formState.errors.customerId && (
-              <span className="text-destructive">{form.formState.errors.customerId.message}</span>
-            )}
-          </FormMessage>
-        </FormItem>
-
-        <FormField
-          control={form.control}
-          name="roomId"
-          render={({ field }) => (
-            <FormItem className=" col-span-2 sm:col-span-1 w-full">
-              <FormLabel>{FORMSTATICS.roomId.label}</FormLabel>
-              <FormControl>
-                <Popover>
-                  <PopoverTrigger asChild>
-                    <FormControl>
-                      <Button
-                        variant="outline"
-                        role="combobox"
-                        className={cn(
-                          "w-full justify-between capitalize truncate",
-                          !field.value && "text-muted-foreground"
-                        )}
-                        disabled={roomOptions.length === 0}
-                      >
-                        {field.value
-                          ? roomOptions.find((room) => room.value === field.value)?.label
-                          : FORMSTATICS.roomId.placeholder}
-                        <ChevronsUpDown className="opacity-50" />
-                      </Button>
-                    </FormControl>
-                  </PopoverTrigger>
-                  <PopoverContent className="w-full p-0">
-                    <Command>
-                      <CommandInput placeholder={FORMSTATICS.roomId.placeholder} className="h-9" />
-                      <CommandList>
-                        <CommandEmpty>No hay habitaciones disponibles</CommandEmpty>
-                        <CommandGroup>
-                          {roomOptions.map((room) => (
-                            <CommandItem
-                              value={room.value}
-                              key={room.value}
-                              onSelect={() => {
-                                form.setValue("roomId", room.value);
-                                const detailedRoom = availableRooms?.find((r) => r.id === room.value);
-                                onRoomSelected(detailedRoom);
-                              }}
-                            >
-                              {room.label}
-                              <Check
-                                className={cn("ml-auto", room.value === field.value ? "opacity-100" : "opacity-0")}
-                              />
-                            </CommandItem>
-                          ))}
-                        </CommandGroup>
-                      </CommandList>
-                    </Command>
-                  </PopoverContent>
-                </Popover>
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
+        <CreateHeaderReservation
+          availableRooms={availableRooms}
+          form={form}
+          onRoomSelected={onRoomSelected}
+          onSearchCustomerFound={onSearchCustomerFound}
+          roomOptions={roomOptions}
         />
 
         <Separator className="col-span-2" />
@@ -288,94 +188,6 @@ export default function CreateReservationForm({
         </div>
 
         <Separator className="col-span-2" />
-
-        {/* <FormField
-          control={form.control}
-          name={"checkInDate"}
-          render={({ field }) => (
-            <FormItem className="flex flex-col gap-2 h-full mt-2">
-              <FormLabel>{FORMSTATICS.checkInDate.placeholder}</FormLabel>
-              <Popover>
-                <PopoverTrigger asChild>
-                  <FormControl>
-                    <Button
-                      variant={"outline"}
-                      className={cn("w-full pl-3 text-left font-normal", !field.value && "text-muted-foreground")}
-                    >
-                      {field.value ? (
-                        // Verifica si es string
-                        typeof field.value === "string" ? (
-                          format(new Date(field.value), "PPP", {
-                            locale: es,
-                          })
-                        ) : null
-                      ) : (
-                        <span>Escoja una fecha</span>
-                      )}
-                      <CalendarIcon className="ml-auto h-4 w-4 opacity-50" />
-                    </Button>
-                  </FormControl>
-                </PopoverTrigger>
-                <PopoverContent className="w-auto p-0" align="start">
-                  <Calendar
-                    mode="single"
-                    selected={typeof field.value === "string" ? new Date(field.value) : undefined}
-                    onSelect={(val) => field.onChange(val?.toLocaleDateString('es-PE', {
-                      timeZone: "America/Lima",
-                    }) ?? "")}
-                    disabled={(date) => date > new Date() || date < new Date("1900-01-01")}
-                    initialFocus
-                  />
-                </PopoverContent>
-              </Popover>
-              <CustomFormDescription required={FORMSTATICS.checkInDate.required} />
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-
-        <FormField
-          control={form.control}
-          name={"checkOutDate"}
-          render={({ field }) => (
-            <FormItem className="flex flex-col gap-2 h-full mt-2">
-              <FormLabel>{FORMSTATICS.checkInDate.placeholder}</FormLabel>
-              <Popover>
-                <PopoverTrigger asChild>
-                  <FormControl>
-                    <Button
-                      variant={"outline"}
-                      className={cn("w-full pl-3 text-left font-normal", !field.value && "text-muted-foreground")}
-                    >
-                      {field.value ? (
-                        // Verifica si es string
-                        typeof field.value === "string" ? (
-                          format(new Date(field.value), "PPP", {
-                            locale: es,
-                          })
-                        ) : null
-                      ) : (
-                        <span>Escoja una fecha</span>
-                      )}
-                      <CalendarIcon className="ml-auto h-4 w-4 opacity-50" />
-                    </Button>
-                  </FormControl>
-                </PopoverTrigger>
-                <PopoverContent className="w-auto p-0" align="start">
-                  <Calendar
-                    mode="single"
-                    selected={typeof field.value === "string" ? new Date(field.value) : undefined}
-                    onSelect={(val) => field.onChange(val?.toISOString() ?? "")}
-                    disabled={(date) => date > new Date() || date < new Date("1900-01-01")}
-                    initialFocus
-                  />
-                </PopoverContent>
-              </Popover>
-              <CustomFormDescription required={FORMSTATICS.checkInDate.required} />
-              <FormMessage />
-            </FormItem>
-          )}
-        /> */}
 
         <FormField
           control={form.control}
@@ -471,271 +283,16 @@ export default function CreateReservationForm({
         )}
 
         {allowGuests && selectedRoom?.RoomTypes?.guests && (
-          <div className="flex flex-col gap-4  sm:col-span-2 animate-ease-in">
-            <FormLabel>{FORMSTATICS.guests.label}</FormLabel>
-            <Table className="w-full overflow-auto">
-              <TableHeader>
-                <TableRow>
-                  <TableHead className="w-[100px]">Nombre y Apell.</TableHead>
-                  <TableHead>Edad</TableHead>
-                  <TableHead>Tipo de Identidad</TableHead>
-                  <TableHead>Nro. de Identidad</TableHead>
-                  <TableHead>Telefono</TableHead>
-                  <TableHead className="text-center">Email</TableHead>
-                  <TableHead className="text-center">Adicional</TableHead>
-                </TableRow>
-              </TableHeader>
-              <TableBody>
-                {controlledFields.map((field, index) => {
-                  // const data = selectedProducts.find((p) => p.id === field.productId);
-                  // const safeData: Partial<OutgoingProducStockForm> = data ?? {};
-                  // const safeWatch = watchFieldArray?.[index] ?? {};
-                  // //const price = safeData.precio ?? 0;
-                  // const safeStorages =
-                  //   safeData.Stock?.map((stock) => ({
-                  //     label: (
-                  //       <div>
-                  //         {stock.Storage.name} {"(Stock "} <span className="text-primary font-bold">{stock.stock}</span>
-                  //         {")"}
-                  //       </div>
-                  //     ),
-                  //     value: stock.Storage.id,
-                  //   })) ?? [];
-
-                  // const stockStorage = safeData.Stock?.find((stock) => stock.Storage.id === safeWatch.storageId) ?? null;
-
-                  // const totalStock = safeData.Stock?.reduce((acc, stock) => acc + stock.stock, 0) ?? 0;
-                  // const dynamicStock = isNaN((stockStorage?.stock ?? 0) - (safeWatch.quantity ?? 0))
-                  //   ? (stockStorage?.stock ?? 0)
-                  //   : (stockStorage?.stock ?? 0) - (safeWatch.quantity ?? 0);
-
-                  // const price = safeData.precio ?? 0;
-                  // const total = isNaN(price * (safeWatch.quantity ?? 0)) ? 0 : price * (safeWatch.quantity ?? 0);
-
-                  const isDocumentTypeSelected = field.documentType !== undefined;
-                  return (
-                    <TableRow key={field.id} className="animate-fade-down duration-500">
-                      <TableCell>
-                        <FormItem>
-                          {/* <div>
-                          <span>{safeData.name ?? "Desconocido"}</span>
-                        </div> */}
-                          <Input {...register(`guests.${index}.name` as const)} className="min-w-[100px] w-full" />
-                          <CustomFormDescription
-                            required={FORMSTATICS.guests.subFields?.name?.required ?? false}
-                            validateOptionalField={true}
-                          ></CustomFormDescription>
-                          <FormMessage>
-                            {form.formState.errors.guests?.[index]?.name &&
-                              form.formState.errors.guests[index]?.name?.message}
-                          </FormMessage>
-                        </FormItem>
-                      </TableCell>
-                      <TableCell>
-                        <FormItem className="min-w-[50px] w-full">
-                          {/* <div>
-                          <span>{safeData.name ?? "Desconocido"}</span>
-                        </div> */}
-                          <Input {...register(`guests.${index}.age` as const)} type="number" min={0} placeholder="0" />
-                          <CustomFormDescription
-                            required={FORMSTATICS.guests.subFields?.age.required ?? false}
-                            validateOptionalField={true}
-                          ></CustomFormDescription>
-                          <FormMessage>
-                            {form.formState.errors.guests?.[index]?.age &&
-                              form.formState.errors.guests[index]?.age?.message}
-                          </FormMessage>
-                        </FormItem>
-                      </TableCell>
-                      <TableCell>
-                        <FormItem>
-                          <Select
-                            {...register(`guests.${index}.documentType` as const)}
-                            defaultValue={field.documentType}
-                            onValueChange={(val) => {
-                              form.setValue(`guests.${index}.documentType`, val as DocumentType);
-                              // calculateProductTotals();
-                            }}
-                          >
-                            <FormControl>
-                              <SelectTrigger className="min-w-[100px] w-full">
-                                <SelectValue
-                                  placeholder={FORMSTATICS.guests.subFields?.documentType.placeholder}
-                                  className="text-ellipsis"
-                                />
-                              </SelectTrigger>
-                            </FormControl>
-                            <SelectContent>
-                              {Object.values(documentTypeStatusConfig).map((documentType, idx) => (
-                                <SelectItem key={`${documentType.value}-${idx}`} value={documentType.value}>
-                                  {documentType.name}
-                                </SelectItem>
-                              ))}
-                            </SelectContent>
-                          </Select>
-                          <CustomFormDescription
-                            required={FORMSTATICS.guests.subFields?.documentType?.required ?? false}
-                            validateOptionalField={true}
-                          ></CustomFormDescription>
-                          <FormMessage>
-                            {form.formState.errors.guests?.[index]?.documentType &&
-                              form.formState.errors.guests[index]?.documentType?.message}
-                          </FormMessage>
-                        </FormItem>
-                      </TableCell>
-                      <TableCell>
-                        <FormItem>
-                          <Input
-                            disabled={!isDocumentTypeSelected}
-                            {...register(`guests.${index}.documentId` as const)}
-                          />
-                          <CustomFormDescription
-                            required={FORMSTATICS.guests.subFields?.documentId.required ?? false}
-                            validateOptionalField={true}
-                          ></CustomFormDescription>
-                          <FormMessage>
-                            {form.formState.errors.guests?.[index]?.documentId &&
-                              form.formState.errors.guests[index]?.documentId?.message}
-                          </FormMessage>
-                        </FormItem>
-                      </TableCell>
-
-                      {/* <TableCell>
-                      <span className="block text-center">
-                        {price.toLocaleString("es-PE", {
-                          style: "currency",
-                          currency: "PEN",
-                        })}
-                      </span>
-                    </TableCell> */}
-                      <TableCell>
-                        {/* <FormItem>
-                          <FormItem>
-                            <FormControl>
-                              <PhoneInput
-                                className="min-w-[170px] w-full"
-                                {...register(`guests.${index}.phone` as const)}
-                                defaultCountry="PE"
-                                placeholder="999 888 777"
-                                value={field.phone}
-                                onChange={(value) => form.setValue(`guests.${index}.phone`, value)}
-                              />
-                            </FormControl>
-                          </FormItem>
-                          <CustomFormDescription
-                            required={FORMSTATICS.guests.subFields?.phone.required ?? false}
-                            validateOptionalField={true}
-                          ></CustomFormDescription>
-                          <FormMessage>
-                            {form.formState.errors.guests?.[index]?.phone &&
-                              form.formState.errors.guests[index]?.phone?.message}
-                          </FormMessage>
-                        </FormItem> */}
-                        <Controller
-                          control={form.control}
-                          name={`guests.${index}.phone`}
-                          render={({ field: { onChange, value } }) => (
-                            <FormItem>
-                              <FormControl>
-                                <PhoneInput
-                                  className="min-w-[170px] w-full"
-                                  defaultCountry="PE"
-                                  placeholder="999 888 777"
-                                  value={value}
-                                  onChange={onChange}
-                                />
-                              </FormControl>
-                              <CustomFormDescription
-                                required={FORMSTATICS.guests.subFields?.phone.required ?? false}
-                                validateOptionalField={true}
-                              ></CustomFormDescription>
-                              <FormMessage />
-                            </FormItem>
-                          )}
-                        />
-                      </TableCell>
-                      <TableCell>
-                        <FormItem>
-                          <Input className="min-w-[100px] w-full" {...register(`guests.${index}.email` as const)} />
-                          <CustomFormDescription
-                            required={FORMSTATICS.guests.subFields?.email.required ?? false}
-                            validateOptionalField={true}
-                          ></CustomFormDescription>
-                          <FormMessage>
-                            {form.formState.errors.guests?.[index]?.email &&
-                              form.formState.errors.guests[index]?.email?.message}
-                          </FormMessage>
-                        </FormItem>
-                      </TableCell>
-                      <TableCell>
-                        <FormItem>
-                          <Textarea
-                            className="min-w-[100px] w-full"
-                            {...register(`guests.${index}.additionalInfo` as const, {
-                              setValueAs: (v) => (v === "" ? undefined : String(v)),
-                            })}
-                          />
-                          <CustomFormDescription
-                            required={FORMSTATICS.guests.subFields?.additionalInfo.required ?? false}
-                            validateOptionalField={true}
-                          ></CustomFormDescription>
-                          <FormMessage>
-                            {form.formState.errors.guests?.[index]?.additionalInfo &&
-                              form.formState.errors.guests[index]?.additionalInfo?.message}
-                          </FormMessage>
-                        </FormItem>
-                      </TableCell>
-                      <TableCell>
-                        <Button
-                          type="button"
-                          variant="outline"
-                          className="hover:bg-destructive hover:text-white"
-                          size="sm"
-                          onClick={() => handleRemoveGuest(index)}
-                        >
-                          <Trash2 />
-                          {/* Eliminar */}
-                        </Button>
-                      </TableCell>
-                    </TableRow>
-                  );
-                })}
-              </TableBody>
-            </Table>
-            <div className="col-span-2 w-full flex flex-col gap-2 justify-center items-center py-4">
-              <Button
-                variant={"outline"}
-                disabled={
-                  selectedRoom?.RoomTypes?.guests && guestNumber >= selectedRoom?.RoomTypes?.guests - 1 ? true : false
-                }
-                type="button"
-                onClick={handleAddGuest}
-                className="flex items-center gap-2"
-              >
-                <ListCheck className="size-4" />
-                Añadir Huésped
-                {selectedRoom?.RoomTypes?.guests && (
-                  <div>
-                    {"("}
-                    <span className="text-primary text-base font-bold">
-                      {/* -1 beacause of the current guest who is making the reservation */}
-                      {selectedRoom?.RoomTypes?.guests - 1 - guestNumber}
-                    </span>{" "}
-                    lugar{selectedRoom?.RoomTypes?.guests > 1 ? "es" : ""} restante
-                    {selectedRoom?.RoomTypes?.guests > 1 ? "s" : ""}
-                    {")"}
-                  </div>
-                )}
-              </Button>
-              <CustomFormDescription
-                required={FORMSTATICS.guests.required}
-                validateOptionalField={true}
-              ></CustomFormDescription>
-              {form.formState.errors.guests && (
-                <FormMessage className="text-destructive">{form.formState.errors.guests.message}</FormMessage>
-              )}
-            </div>
-          </div>
+          <CreateReservationGuestTable
+            controlledFieldArray={controlledFieldArray}
+            guestNumber={guestNumber}
+            handleAddGuest={handleAddGuest}
+            handleRemoveGuest={handleRemoveGuest}
+            form={form}
+            register={register}
+            selectedRoom={selectedRoom}
+            watch={watch}
+          />
         )}
 
         <Separator className="col-span-2" />
